// src/api/models/FormatEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `multiple_choice` - Multiple Choice
 * * `ox_selection` - O/X selection
 * * `text_input` - Short Answer
 * * `essay` - Essay
 */
export enum FormatEnum {
  MULTIPLE_CHOICE = 'multiple_choice',
  OX_SELECTION = 'ox_selection',
  TEXT_INPUT = 'text_input',
  ESSAY = 'essay',
}